Move linalgext external models
